% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ecdf_cpp}
\alias{ecdf_cpp}
\title{C++ implementation of empirical cdf}
\usage{
ecdf_cpp(sobs, sref)
}
\arguments{
\item{sobs}{the observation for which the cdf is to be evaluated}

\item{sref}{the reference observations from which the cdf is to be constructed}
}
\description{
C++ implementation of empirical cdf
}
\examples{
library(microbenchmark)
fi <- rnorm(n=1000)
n_new <- 100
fnew <- rnorm(n=n_new) 
microbenchmark(#sum.I(c(fi, fnew), FUN=">=", fnew)/n_new,
   ecdf(fnew)(c(fi, fnew)),
   ecdf_cpp(c(fi, fnew), sort(fnew))/n_new
)
}

