% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(p)
}
\arguments{
\item{p}{a number between \code{0} and \code{1} (e.g. a probability)}
}
\value{
a real number between \code{-Inf} and \code{Inf}
}
\description{
logit function
}
