% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sslcov_test.R
\name{print.ssl_test}
\alias{print.ssl_test}
\title{Print methods for ssl_test results}
\usage{
\method{print}{ssl_test}(x, ...)
}
\arguments{
\item{x}{a \code{ssl_test} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print methods for ssl_test results
}

