% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat_cond.R
\name{rhat_cond}
\alias{rhat_cond}
\title{Semi-supervised correlation estimation}
\usage{
rhat_cond(data, nn, outcome_name = NULL, covariate_name = NULL,
  surrogate_name = NULL, bw = NULL, cdf_trans = TRUE, weights = NULL,
  adjust_covariates_name = NULL, do_interact = TRUE)
}
\arguments{
\item{data}{the data. The first \code{nn} rows should be the labeled data, the
remaining rows should be the unlabeled data.}

\item{nn}{the number of labeled data}

\item{outcome_name}{a character string containing the name of the
column from data containing the partly missing outcome of interest}

\item{covariate_name}{a character string containing the name of the
column from data containing the covariate to be related to the outcome of
interest}

\item{surrogate_name}{a character string vector containing the name of the
column(s) from data containing the surrogate variable(s)}

\item{bw}{the bandwidth to use}

\item{cdf_trans}{a logical flag indicating wether the smoothing should be performed on the data
transformed with their cdf. Default is TRUE. See Details.}

\item{weights}{a vector of weights in case a weighted version of the
correlation has to be computed. Default is \code{NULL}, in which case, no
weighting is done.}

\item{adjust_covariates_name}{optional vector of names of the covariates to adjust on during imputation and smoothing.
Default is \code{NULL}.}

\item{do_interact}{logical flag indicating whether interactins between \code{x} and
covariates should be taken into account when imputing \code{y}. Default is \code{TRUE}.}
}
\value{
a list with the following elements:\itemize{
   \item rhat
   \item bw the bandwith used
   \item data_sup
   \item W_unlabel
   }
}
\description{
This function estimates the correlation between an outcome available only
for a small subset of the data and a covariate. The outcome is imputed to all
the data using a smoothed predictor learned thanks to a set of surrogate variables,
available for all the data.
}
\details{
Smoothing over the CDF transformed data prevents some tail estimation issues when the new data are subsequently large.
}
\seealso{
\code{\link{smooth_ssl}}
}

