% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_Boris_1runFn.R
\name{simu_Boris_1runFn}
\alias{simu_Boris_1runFn}
\title{Simulation function computing 1 run implemented by Boris}
\usage{
simu_Boris_1runFn(NN = 2e+05, nn_divide = 1000, nperturb = 500,
  mySigma = matrix(rep(0.3, 16), 4, 4) + 0.7 * diag(4), underTheNull = TRUE,
  verbose = FALSE, i = NULL, extreme = TRUE, beta_2bePtb = TRUE)
}
\description{
Simulation function computing 1 run implemented by Boris
}
\examples{

sims1 <- simu_Boris_1runFn(nperturb=2)

}

