% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_sslCPP.R
\name{smooth_sslCPP}
\alias{smooth_sslCPP}
\title{Computes a smoothed projection}
\usage{
smooth_sslCPP(ri, fi, fnew, rsup, wgt = NULL, bw = NULL, cdf_trans = TRUE,
  condi = FALSE)
}
\arguments{
\item{ri}{label data correlation}

\item{fi}{the odl data}

\item{fnew}{the new data to be}

\item{wgt}{facultative weights (used for perturbations)}

\item{bw}{bandwith}

\item{cdf_trans}{a logical flag indicating wether the smoothing should be
performed on the data transformed with their cdf. Default is \code{TRUE}.
See Details.}
}
\value{
a
}
\description{
Smoothing predictor using Gaussian kernel
}
\details{
Smoothing over the CDF transformed data preven_learns some tail estimation issues
when the new data are subsequen_learnly large.
}

