% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sslcov-package.R
\docType{package}
\name{sslcov-package}
\alias{sslcov-package}
\alias{sslcov}
\title{ssl}
\description{
This package implelments
}
\details{
\tabular{ll}{
Package: \tab sslcov\cr
Type: \tab Package\cr
Version: \tab sslcov.0.0.5\cr
Date: \tab 2016-07-25\cr
License: \tab \href{http://www.gnu.org/licenses/lgpl.txt}{MIT}\cr
}
The main function
}
\author{
Boris P. Hejblum, Stephanie Chan, Tianxi Cai
--- Maintainer: Stephanie Chan
}
