% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.I.R
\name{sum.I}
\alias{sum.I}
\title{Cumulative indicator computation ?}
\usage{
\method{sum}{I}(yy, FUN, Yi, Vi = NULL)
}
\value{
a vector of length(yy)
}
\description{
Cumulative indicator computation ?
}
\keyword{internal}
