% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sslcov_test.R
\name{summary.ssl_test}
\alias{summary.ssl_test}
\title{Print method for ssl_test results}
\usage{
\method{summary}{ssl_test}(object, ...)
}
\arguments{
\item{object}{a \code{ssl_test} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print method for ssl_test results
}
