% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnormC_multi}
\alias{dnormC_multi}
\title{C++ implementation of univariate Normal of mean zero pdf for multiple inputs}
\usage{
dnormC_multi(x, m, sd, Log = FALSE)
}
\arguments{
\item{x}{data matrix of dimension p x n, p being the dimension of the
data and n the number of data points}

\item{sd}{the standard deviation}

\item{logical}{flag for returning the log of the probability density
function. Defaults is \code{TRUE}}
}
\value{
matrix of densities of dimension \code{p x n}.
}
\description{
C++ implementation of univariate Normal of mean zero pdf for multiple inputs
}

