% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulating data}
\usage{
sim_data(ntot, missing = FALSE, incorrect = FALSE, b_G = 0.6,
  Sigma = diag(4))
}
\arguments{
\item{ntot}{the number of observation to simulate}

\item{missing}{logical flag wether
. Default is FALSE.}

\item{b_G}{strength parameter for linear association between the covariate "G"
and the outcome of interest "Y".}
}
\value{
simulated data matrix of size \code{ntot x 4}
}
\description{
Simulating data
}
\keyword{internal}

