% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_ssl.R
\name{smooth_ssl}
\alias{smooth_ssl}
\title{Computes a smoothed projection}
\usage{
smooth_ssl(ri, fi, fnew, wgti = NULL, bw = NULL, cdf_trans = TRUE, rsup)
}
\arguments{
\item{ri}{label data correlation}

\item{fi}{the labeled data}

\item{fnew}{the new data to be predicted}

\item{bw}{bandwith}

\item{cdf_trans}{a logical flag indicating wether the smoothing should be
performed on the data transformed with their cdf. Default is \code{TRUE}.
See Details.}

\item{wgt}{facultative weights (used for perturbations)}
}
\value{
a
}
\description{
Smoothing predictor using Gaussian kernel
}
\details{
Smoothing over the CDF transformed data preven_learns some tail estimation issues
when the new data are subsequen_learnly large.
}

