% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sslcov_test.R
\name{sslcov_test}
\alias{sslcov_test}
\title{Wrapper function for testing TODO}
\usage{
sslcov_test(y, x, index_sup, surrogate, adjust_covariates = NULL,
  sampling_weights = rep(1/length(y), length(index_sup)), nperturb = 500,
  do_interact = TRUE, condi = FALSE, ptb = TRUE)
}
\arguments{
\item{y}{vector of length \code{N} of the outcome}

\item{x}{vector of length \code{N} containing the covariate of interest}

\item{index_sup}{vector of length \code{n} indicating which observations whave supervised info}

\item{surrogate}{matrix with \code{N} rows containing the surrogate information}

\item{adjust_covariates}{optional matrix with \code{N} rows containing covariates to adjust on in
the semi-supervised estimation. Default is \code{NULL}}

\item{sampling_weights}{optional vector of length \code{n} containing the weights with which the
\code{n} supervised observation were sampled from the population. Default is random sampling (all the weigths equal to 1).}

\item{nperturb}{The number of perturbations to be run. Default is \code{500}.}

\item{do_interact}{logical flag indicating whether interactins between \code{x} and
covariates should be taken into account when imputing \code{y}. Default is \code{TRUE}.}

\item{condi}{logical flag indicating whether the covariance estimated should be condition on the
covariates indicated in . Default is \code{TRUE}.}

\item{ptb}{logical flag indicating whether to use perturbation to calculate the variance instead of using the asymptotic variance}
}
\description{
Wrapper function for testing TODO
}
\examples{
\dontrun{
#rm(list=ls())

#Simulate data
nn_divide <- 10
NN <- 2000
nn <- NN/nn_divide
mySigma <- matrix(rep(0.3,16), 4, 4) + 0.7*diag(4)
beta <- 0.6 #0
set.seed(1234)
data_sim <- sim_data(ntot = NN, Sigma = mySigma, b_G = beta, cond_cov = TRUE)
cov_sim <- data_sim$cov_cond
data_sim <- data_sim$data
cov_sim
cov(data_sim[,"Y"], data_sim[,"G"])
es <- extremeSampling(data_sim, nn=nn, surrogate_name=c("S1", "S2", "S3"))
data_sampled <- rbind(data_sim[es$extreme_index,], data_sim[-es$extreme_index,])

#True Covariance:
cov(data_sim[,"Y"], log(1+data_sim[,"G"]))

res_ssl_randomsampling <- sslcov_test(y = data_sim[,"Y"], x = log(1 + data_sim[,"G"]),
                                     index_sup = 1:nn,
                                     surrogate = data_sim[,c("S1", "S2", "S3")],
                                     do_interact=FALSE, condi = FALSE, ptb=FALSE)
res_ssl <- sslcov_test(y = data_sampled[,"Y"], x = log(1 + data_sampled[,"G"]),
                      index_sup = 1:nn,
                      surrogate = data_sampled[,c("S1", "S2", "S3")],
                      sampling_weights = es$weights,
                      do_interact=FALSE, condi = FALSE, ptb=FALSE)
res_ssl_noWeights <- sslcov_test(y = data_sampled[,"Y"], x = log(1 + data_sampled[,"G"]),
                                index_sup = 1:nn,
                                surrogate = data_sampled[,c("S1", "S2", "S3")],
                                do_interact=FALSE, condi = FALSE, ptb=FALSE)


# Conditional:

cov(data_sim[,"Y"], data_sim[,"G"])
cov(lm(data_sim[,"Y"]~data_sim[,"X"])$residuals, data_sim[,"G"] -
   exp(MASS::glm.nb(data_sim[,"G"]~data_sim[,"X"])$linear.predictors))
#library(profvis)
#profvis(
res_ssl_random_condi <- sslcov_test(y = data_sim[,"Y"], x = data_sim[,"G"], index_sup = 1:nn,
                             surrogate = data_sim[,c("S1", "S2", "S3")],
                             adjust_covariates = data_sim[,"X", drop=FALSE],
                             do_interact=FALSE, condi = TRUE, ptb=FALSE)
#)
res_ssl_condi <- sslcov_test(y = data_sampled[,"Y"], x = data_sampled[,"G"], index_sup = 1:nn,
                         surrogate = data_sampled[,c("S1", "S2", "S3")],
                         adjust_covariates = data_sampled[,"X", drop=FALSE],
                         sampling_weights = es$weights,
                         do_interact=FALSE, condi = TRUE, ptb=FALSE)
#
res_ssl_noWeights_condi <- sslcov_test(y = data_sampled[,"Y"], x = data_sampled[,"G"],
                             index_sup = 1:nn, surrogate = data_sampled[,c("S1", "S2", "S3")],
                             adjust_covariates = data_sampled[,"X", drop=FALSE],
                             do_interact=FALSE, condi = TRUE, ptb=FALSE)
}

}
